+++ .changeset/healthy-rats-confess.md
@@ -3,14 +3,103 @@
 '@shopify/create-hydrogen': patch
 ---
 
-Use type `HydrogenContext` for all the env.d.ts
+1. Create a app/lib/context file and use `createHydrogenContext` in it.
+
+```.ts
+// in app/lib/context
+
+import {createHydrogenContext} from '@shopify/hydrogen';
+
+export type CustomAppLoadContext = Awaited<ReturnType<typeof createAppLoadContext>>;
+
+export async function createAppLoadContext(
+  request: Request,
+  env: Env,
+  executionContext: ExecutionContext,
+) {
+    const hydrogenContext = createHydrogenContext({
+      env,
+      request,
+      cache,
+      waitUntil,
+      session,
+      i18n: {language: 'EN', country: 'US'},
+      cart: {
+        queryFragment: CART_QUERY_FRAGMENT,
+      },
+      // ensure to overwrite any options that is not using the default values from your server.ts
+    });
+
+  return hydrogenContext;
+}
+
+```
+
+2. Use `createAppLoadContext` method in server.ts Ensure to overwrite any options that is not using the default values.
+
+```diff
+// in server.ts
+
+- import {
+-   createCartHandler,
+-   createStorefrontClient,
+-   createCustomerAccountClient,
+- } from '@shopify/hydrogen';
++ import {createAppLoadContext} from '~/lib/context';
+
+export default {
+  async fetch(
+    request: Request,
+    env: Env,
+    executionContext: ExecutionContext,
+  ): Promise<Response> {
+
+-   const {storefront} = createStorefrontClient(
+-     ...
+-   );
+
+-   const customerAccount = createCustomerAccountClient(
+-     ...
+-   );
+
+-   const cart = createCartHandler(
+-     ...
+-   );
+
++   const appLoadContext = await createAppLoadContext(
++      request,
++      env,
++      executionContext,
++   );
+
+    /**
+      * Create a Remix request handler and pass
+      * Hydrogen's Storefront client to the loader context.
+      */
+    const handleRequest = createRequestHandler({
+      build: remixBuild,
+      mode: process.env.NODE_ENV,
+-      getLoadContext: (): AppLoadContext => ({
+-        session,
+-        storefront,
+-        customerAccount,
+-        cart,
+-        env,
+-        waitUntil,
+-      }),
++      getLoadContext: () => appLoadContext,
+    });
+  }
+```
+
+3. Use infer type for AppLoadContext in env.d.ts
 
 ```diff
 // in env.d.ts
 
-+ import type {HydrogenContext} from '@shopify/hydrogen';
++ import type {CustomAppLoadContext} from '~/lib/context';
 
-+ interface AppLoadContext extends HydrogenContext<AppSession> {
++ interface AppLoadContext extends CustomAppLoadContext {
 - interface AppLoadContext {
 -  env: Env;
 -  cart: HydrogenCart;
