--- examples/custom-cart-method/env.d.ts
@@ -1,48 +1,0 @@
-/// <reference types="vite/client" />
-/// <reference types="@shopify/remix-oxygen" />
-/// <reference types="@shopify/oxygen-workers-types" />
-
-// Enhance TypeScript's built-in typings.
-import '@total-typescript/ts-reset';
-
-import type {
-  HydrogenContext,
-  HydrogenSessionData,
-  HydrogenEnv,
-} from '@shopify/hydrogen';
-import type {AppSession} from '~/lib/session';
-
-declare global {
-  /**
-   * A global `process` object is only available during build to access NODE_ENV.
-   */
-  const process: {env: {NODE_ENV: 'production' | 'development'}};
-
-  interface Env extends HydrogenEnv {
-    // declare additional Env parameter use in the fetch handler and Remix loader context here
-  }
-}
-
-declare module '@shopify/remix-oxygen' {
-  interface AppLoadContext
-    extends HydrogenContext<
-      AppSession,
-      /***********************************************/
-      /**********  EXAMPLE UPDATE STARTS  ************/
-      {
-        updateLineByOptions: (
-          productId: string,
-          selectedOptions: SelectedOptionInput[],
-          line: CartLineUpdateInput,
-        ) => Promise<CartQueryDataReturn>;
-      }
-      /**********   EXAMPLE UPDATE END   ************/
-      /***********************************************/
-    > {
-    // declare additional Remix loader context here
-  }
-
-  interface SessionData extends HydrogenSessionData {
-    // declare local additions to the Remix session data here
-  }
-}
