+++ examples/third-party-queries-caching/tsconfig.json
@@ -1,6 +1,11 @@
 {
   "extends": "../../templates/skeleton/tsconfig.json",
-  "include": ["./**/*.d.ts", "./**/*.ts", "./**/*.tsx"],
+  "include": [
+    "./**/*.d.ts",
+    "./**/*.ts",
+    "./**/*.tsx",
+    "../../templates/skeleton/*.d.ts"
+  ],
   "compilerOptions": {
     "baseUrl": ".",
     "paths": {
